chore: address PR comments, add saveAllIssuers helper
