+++ packages/zoe/src/contractSupport/index.js
@@ -21,4 +21,5 @@ export {
   swapExact,
   depositToSeat,
   withdrawFromSeat,
+  saveAllIssuers,
 } from './zoeHelpers';
